+++ CHANGELOG.md
@@ -3,6 +3,11 @@ autossh-init Cookbook CHANGELOG
 This file is used to list changes made in each version of the
 autossh_init cookbook.
 
+v0.1.3
+------
+- Improve README.md
+- Activate 'ssh_hosts' attributes.
+
 v0.1.2
 ------
 - Change name to 'autossh-init'